# infra/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ResearchMate API
  description: Local-first ResearchMate API with swappable cloud/provider adapters.
  version: 0.1.0
paths:
...
  /api/v1/quiz:
    post:
      tags:
      - quiz
      summary: Create Quiz
      operationId: create_quiz_api_v1_quiz_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/projects/{project_id}/quiz:
    get:
      tags:
      - quiz
      summary: List Quiz History
      operationId: list_quiz_history_api_v1_projects__project_id__quiz_get
      security:
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{run_id}/sources:
    get:
      tags:
      - sources
      summary: Get Run Sources
      operationId: get_run_sources_api_v1_runs__run_id__sources_get
      security:
      - bearerAuth: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSourcesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Citation:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        source_type:
          $ref: '#/components/schemas/SourceType'
        document_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Document Id
        chunk_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Chunk Id
        page_no:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Page No
        slide_no:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Slide No
        url:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Url
        quote:
          type: string
          maxLength: 1200
          minLength: 1
          title: Quote
        claim_id:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Claim Id
      type: object
      required:
      - id
      - source_type
      - quote
      title: Citation
...
    Difficulty:
      type: string
      enum:
      - easy
      - medium
      - hard
      title: Difficulty
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QuizHistoryResponse:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        quiz_sets:
          items:
            $ref: '#/components/schemas/QuizSet'
          type: array
          maxItems: 100
          title: Quiz Sets
      type: object
      required:
      - project_id
      title: QuizHistoryResponse
    QuizQuestion:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        type:
          type: string
          enum:
          - single_choice
          - short_answer
          title: Type
        question:
          type: string
          maxLength: 1200
          minLength: 1
          title: Question
        options:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 4
          - type: 'null'
          title: Options
        answer:
          type: string
          maxLength: 1200
          minLength: 1
          title: Answer
        explanation:
          type: string
          maxLength: 2000
          minLength: 1
          title: Explanation
        difficulty:
          $ref: '#/components/schemas/Difficulty'
          default: medium
        source_citations:
          items:
            $ref: '#/components/schemas/Citation'
          type: array
          maxItems: 12
          title: Source Citations
      type: object
      required:
      - id
      - type
      - question
      - answer
      - explanation
      title: QuizQuestion
    QuizRequest:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        prompt:
          type: string
          maxLength: 4000
          minLength: 1
          title: Prompt
          default: /quiz
        selected_mode:
          $ref: '#/components/schemas/SourceMode'
          default: local_only
        single_choice_count:
          type: integer
          maximum: 20.0
          minimum: 0.0
          title: Single Choice Count
          default: 5
        short_answer_count:
          type: integer
          maximum: 20.0
          minimum: 0.0
          title: Short Answer Count
          default: 3
      type: object
      required:
      - project_id
      title: QuizRequest
    QuizResponse:
      properties:
        quiz_set:
          $ref: '#/components/schemas/QuizSet'
        run_id:
          type: string
          format: uuid
          title: Run Id
        trace_id:
          type: string
          format: uuid
          title: Trace Id
        validation_status:
          type: string
          enum:
          - passed
          - failed
          - retrying
          title: Validation Status
      type: object
      required:
      - quiz_set
      - run_id
      - trace_id
      - validation_status
      title: QuizResponse
    QuizSet:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        mode:
          $ref: '#/components/schemas/SourceMode'
        sources:
          $ref: '#/components/schemas/SourceSummary'
        questions:
          items:
            $ref: '#/components/schemas/QuizQuestion'
          type: array
          maxItems: 40
          minItems: 1
          title: Questions
      type: object
      required:
      - id
      - mode
      - sources
      - questions
      title: QuizSet
    RunSourcesResponse:
      properties:
        run_id:
          type: string
          format: uuid
          title: Run Id
        summary:
          $ref: '#/components/schemas/SourceSummary'
        citations:
          items:
            $ref: '#/components/schemas/Citation'
          type: array
          maxItems: 120
          title: Citations
      type: object
      required:
      - run_id
      - summary
      title: RunSourcesResponse
    SourceMode:
      type: string
      enum:
      - auto
      - local_only
      - web_only
      - hybrid
      title: SourceMode
    SourceSummary:
      properties:
        local_chunks:
          type: integer
          maximum: 50.0
          minimum: 0.0
          title: Local Chunks
          default: 0
        web_pages:
          type: integer
          maximum: 20.0
          minimum: 0.0
          title: Web Pages
          default: 0
      type: object
      title: SourceSummary
    SourceType:
      type: string
      enum:
      - local_doc
      - web_page
      title: SourceType
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorResponse:
      $defs:
        ErrorDetail:
          properties:
            code:
              maxLength: 80
              minLength: 2
              title: Code
              type: string
            message:
              maxLength: 300
              minLength: 1
              title: Message
              type: string
            request_id:
              maxLength: 120
              minLength: 4
              title: Request Id
              type: string
          required:
          - code
          - message
          - request_id
          title: ErrorDetail
          type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
      title: ErrorResponse
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or local dev token